requirements.txt
openvino==2024.1.0
openvino-dev==2024.1.0
optimum-intel[openvino]
transformers
h5py
numpy
scikit-learn
streamlit
plotly
Pillow
matplotlib